refactor(cli): organize project commands
