+++ app/src/cli/commands/project/mod.rs
@@ -1,3 +1,6 @@
+mod args;
 mod handler;
 
+pub use args::ProjectCommand;
+pub(crate) use args::HELP;
 pub(super) use handler::execute;
